fix(app): encode request values in batch, manifest, facility, dashboard, and admin queries (#142)
Extends the module-report treatment to the shared surfaces: every request
value in the batch, specimen referral manifest, facility, dashboard
drilldown, and admin monitoring endpoints now passes an encoder at the
clause site - (int) casts for single ids, DatabaseService::inIntList() for
IN () lists, escape() for text, escapeLike() for LIKE patterns - and the
remaining hand-rolled DataTables search and sort blocks are replaced with
the shared escaped helpers. Twelve endpoints join the
check-sql-value-safety invariant, which now covers 82.

The batch and manifest list endpoints route the test-type slug through
TestsService::getTestTableName(), which throws on an unknown type, so a
single validated variable now feeds the queries and the action links
instead of the raw request value.

Defects fixed on sight along the way:

- Two path traversals: the facility logo and user signature removal
  handlers passed a request-supplied filename straight into the delete
  path, so a crafted ../ sequence could delete files outside the upload
  folder; basename() pins both deletes inside their own directories.
- The dashboard status drilldown page interpolated the base64-decoded
  status id into its query and into inline JS; an int cast at the source
  closes both. Its list endpoint also crashed on a filterless request
  because a string fallback for the WHERE clause was later appended to as
  an array; the clause now stays an array throughout.
- getFacilityDetails relied on the including page for its database handle
  and never sanitized its input; it now bootstraps itself like every
  other endpoint.

The api-dashboard duplicates detail endpoint already binds its search value
as a parameter and needed no change. The import processors and reference
save helpers only concatenate request values into activity-log text and
bound array updates, not SQL.

Verified against the local DRC copy: the drilldown predicate probe returns
identical counts old versus new, and every rewritten query shape runs.

+++ app/dashboard/vlTestResultStatus.php
@@ -12,7 +12,9 @@ require_once APPLICATION_PATH . '/header.php';
 /** @var Psr\Http\Message\ServerRequestInterface $request */
 $request = AppRegistry::get('request');
 $_GET = _sanitizeInput($request->getQueryParams());
-$id = (isset($_GET['id'])) ? base64_decode((string) $_GET['id']) : null;
+// A status id is numeric; the cast keeps a crafted encoded value out of the
+// query below and out of the inline JS this page prints the id into.
+$id = (isset($_GET['id'])) ? (int) base64_decode((string) $_GET['id']) : null;
 
 $date = base64_decode((string) $_GET['d']);
 $tsQuery = "SELECT status_name FROM r_sample_status WHERE status_id = '" . $id . "'";
